requirements.txt
Pillow==10.2.0
PyAutoGUI==0.9.54 